Pin timezonefinder, consolidate utilities, rename MHKConfig, removed defunct conda build (#500)
* Removing timezonefinder

* Consolidate utilities; rename MHKWaveConfig

* Pinned timezonefinder

* Removed defunct conda build

* Fixed import
