--- .github/workflows/conda_build.yml
@@ -1,24 +1,0 @@
-name: Conda Build and Upload
-
-on:
-  release:
-    types: [published]
-
-jobs:
-  build:
-    name: Conda
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v4
-    - uses: conda-incubator/setup-miniconda@v3
-      with:
-        auto-update-conda: true
-        python-version: 3.11
-    - name: Build and upload conda package
-      shell: bash -l {0}
-      env:
-        ANACONDA_TOKEN: ${{ secrets.ANACONDA_TOKEN }}
-      run: |
-        conda install --yes --quiet conda-build conda-verify anaconda-client
-        conda build conda.recipe/ -c nrel -c conda-forge -c sunpower
-        anaconda -t $ANACONDA_TOKEN upload -u nrel $(conda build conda.recipe/ -c nrel -c conda-forge -c sunpower --output)
