+++ examples/legacy/analysis/single_location.py
@@ -24,6 +24,7 @@ from hopp.simulation.technologies.sites.flatirons_site import flatirons_site as 
 from hopp.simulation.hybrid_simulation import HybridSimulation
 from hopp.tools.analysis import create_cost_calculator
 from hopp.tools.resource import *
+from hopp.tools.resource.resource_tools import get_offset
 from hopp import ROOT_DIR
 
 resource_dir = ROOT_DIR / "simulation" / "resource_files"
