+++ hopp/simulation/technologies/layout/pv_inverter.py
@@ -2,7 +2,7 @@ from typing import Union
 import PySAM.Pvsamv1 as pv_detailed
 import PySAM.Pvwattsv8 as pv_simple
 
-from hopp.tools.utils import flatten_dict
+from hopp.utilities import flatten_dict
 
 def get_inverter_attribs(model: Union[pv_simple.Pvwattsv8, pv_detailed.Pvsamv1, dict], only_ref_values=True) -> dict:
     """
