+++ hopp/simulation/technologies/layout/shadow_flicker.py
@@ -8,9 +8,9 @@ from shapely.affinity import translate
 from shapely.geometry import Point
 from shapely.geometry import Polygon, MultiPolygon, MultiPoint
 from shapely.ops import unary_union
-import timezonefinder
 from pysolar.solar import *
 from pvmismatch import *
+import timezonefinder
 
 from hopp.simulation.technologies.layout.pv_module import *
