+++ hopp/simulation/technologies/power_source.py
@@ -7,7 +7,7 @@ import PySAM.Singleowner as Singleowner
 from hopp.simulation.technologies.sites.site_info import SiteInfo
 from hopp.utilities.log import hybrid_logger as logger
 from hopp.simulation.technologies.dispatch.power_sources.power_source_dispatch import PowerSourceDispatch
-from hopp.tools.utils import array_not_scalar, equal
+from hopp.utilities import array_not_scalar, equal
 from hopp.utilities.log import hybrid_logger as logger
 from hopp.simulation.base import BaseClass
