+++ hopp/tools/resource/__init__.py
@@ -1,2 +1,2 @@
-from .resource_tools import get_country, filter_sites, get_offset, extrapolate_wind_speed
+from .resource_tools import get_country, filter_sites, extrapolate_wind_speed
 from .resource_loader.resource_loader_files import resource_loader_file
