+++ pyproject.toml
@@ -42,8 +42,7 @@ dependencies = [
     "scipy",
     "shapely>=2",
     "setuptools",
-    "timezonefinder",
-    "urllib3",
+    "timezonefinder==6.5.9",
     "openpyxl",
     "attrs",
     "utm",
